Add a pass to drop `shape_invariant` attribute from while ops only within device cluster.
This is similar to the `tf-drop-while-shape-invariant` pass run in bridge phase 2, except that it is restricted to run only within tf_device.cluster ops. Assuming that the cluster ops are going to be compiled to XLA, which only supports While ops where the operand and result shape matches, we drop the attribute to enable more exact shapes. More precise shapes allow better canonicalization of WhileRegion ops necessary for resource op lifting.

PiperOrigin-RevId: 360262325
Change-Id: Id51275e3c9aacfaa1017001f0ae880973fe90e2a