Pass output sharding of all outputs to `MapClusterOutputIndexWithRegionOutputIndex`.
Output sharding of all prior outputs is required to correctly compute the mapping of cluster output and per core output. The previous logic of `GetTileShardedOutputsToMerge` and `HandleTileShardedOutputs` was not correct as it only passed the output sharding of one output.
This is exposed when one of the earlier outputs is placed on a single core ("MAXIMAL" sharding) and the latter output is split between cores ("OTHER" i.e., tiled sharding).

PiperOrigin-RevId: 361204985
Change-Id: I215f07147910f6fbaee4b522801360b924781cdd