Fix bug when LossScaleOptimizer and gradient_aggregator were used.
Before, if there were multiple processes each with a LossScaleOptimizer, and the inner optimizer had a custom gradient_aggregator to aggregate gradients across the processes, the LossScaleOptimizers in the processes could become out of sync. In particular, if a strict subset of replicas encountered NaN gradients, only the LossScaleOptimizers with NaN gradients would skip the step, while the LossScaleOptimizers with finite gradients would still apply gradients

Resolves https://github.com/tensorflow/tensorflow/issues/46863. Horovod was broken with LossScaleOptimizer in TF 2.4 due to this bug, but they patched in a workaround. With this fix, the workaround will not be necessary in TF 2.5.

The issue was caused by the fact the gradients were checked for NaNs before they were all-reduced. This was handled correctly with tf.distribute.Strategy but not with an arbitrary gradient_aggregator. Now, the gradients are all-reduced before they are checked for NaNs, which means if any replica has NaNs, they all will have NaNs.

This potentially will allow better performance in the future as well, as it allows earlier-computed gradients to be all-reduced in parallel with the computation of later-computed gradients.

PiperOrigin-RevId: 361236948
Change-Id: I252df898b71994a2c58543207db96abdea6f2835