Introduce a `RefineTypeWith` method in shape_inference (NFC)
This method takes two TensorFlow types and tries to produce a more refined
type, where "refine" implies that more static information are known about
it.
This patch is NFC, the new method is introduced and correctness checked by
changing the implementation of `CanRefineTypeWith` to use this new method.
A future patch will refactor the uses of CanRefineTypeWith to use this in
more places directly, and in particular to improve our handling of
types with subtypes (resource and variant).

PiperOrigin-RevId: 361458355
Change-Id: I2960c8a9eefb6fa0d1dfbabc645ad7c390462a16