[XLA:Python] Add a PyBuffer::Clone() method that clones the Python wrapper object for a PjRtBuffer.
Changes ownership of PjRtBuffers in PyBuffer to shared ownership.

JAX's C++ jit keeps certain metadata on PyBuffer objects, but we need to be able to create copies of buffers with different metadata without copying the buffers. The canonical example of this is the sticky device property of JAX arrays. Currently JAX does this by wrapping a Python _DeviceArray object around the C++ buffer object, but we would like to remove the Python wrapper.

PiperOrigin-RevId: 362128416
Change-Id: Ia559655d0d0dad29a1bf84fb07c7fd6b792d0fd5