[PJRT] Remove `PjRtBuffer::donation_semaphore_`.
We can save a mutex acquisition in the common case by using the PjRtBuffer's mutex for waiting for outstanding donations, rather than a separate mutex. The Semaphore implementation is implemented as mu.Await() anyway, and it doesn't greatly matter which mutex we wait on. We use the holds_ field to detect the number of outstanding holds.

PiperOrigin-RevId: 363237344
Change-Id: Ie0799b4e20362d895e4f98e04a1082dd96259931