Remove guards around b/143684500.
Several kernels have previously been disabled for std::complex due
to lack of operator support on device.  This is no longer the case,
and these kernels can now be enabled.

NOTE: we also needed to replace `a+=b` and `a-=b` with `a=a+b` and
`a=a-b` for use with `std::complex<T>` on device.  Unfortunately MSVC
does not stricly follow the c++ standard, and unnecessarily specializes
these operators. This prevents us from using Eigen's device-compatible
overrides.  Without these overrides, the original compound assignments
`+=` and `-=` silently fail on GPU.
PiperOrigin-RevId: 363775580
Change-Id: I6aa37dfd07f066c32e4a763917e851af00cc9aab