[TF:TRT] TensorRT execution context should not retain GPU memory.
Previously, when we create TensorRT execution context, we also allocate the GPU
memory the execution context requires in order to execute the TensorRT engine
associated with the execution context. This GPU memory won't be released until
the execution context is destroyed when the containing TRTEngineOp is
destructed. A TF-TRT converted graph often has multiple execution contexts, due
to multiple TRTEngineOps or multiple execution contexts for a single
TRTEngineOp. For this reason, retaining GPU memory in each execution context
during the graph execution can cause unnecessary excessive GPU memory usage.

This CL delays the allocation of GPU memory for an execution context until the
execution context is used to enqueue a TensorRT engine for execution and
releases the GPU memory when the execution context finishes executing the
TensorRT engine.

This CL leaves unused fields or unused parameters in some data structures or
subroutines. Cleaning up the code will be done in a follow up CL.

PiperOrigin-RevId: 363890276
Change-Id: Ib91ddf104417b247d4d112fa5052893dbf42e215