[TF:TRT] Code cleanup.
Remove the memory allocator and device memory fields from the ExecutionContext
as it no longer needs to manage the device memory used for the execution.
Remove the memory allocator from the routines for creating ExecutionContext.

Also replace a non-const reference parameter with a pointer parameter in
CreateExecutionContexts and rename a field and some parameters in class
EngineContext.

Fix affected tests.

PiperOrigin-RevId: 363962954
Change-Id: I77754754c38b46c7091e01c71d11efbd24761484