Change the keras saved_model metadata to use the copied version data from keras.
This is trying to isolate the dependency from keras to TF so that OSS keras could
rely on the PIP package only (the proto build deps is not available from PIP).

The version information is current populated with dummy data, and not being consumed by the Keras code. Remove the old field and replace with new version should have no impact for any existing saved saved_model binaries.

PiperOrigin-RevId: 364339229
Change-Id: I5e8be0b4c88aa770365c2566cf4ecac0101bfde8