Add pre-load CPU feature guard check
Compile a small shared library that performs an x86 CPU feature guard
check that compares the features enabled during compile-time with the
features available at run-time (using the `CPUID` instruction). This
comparison is done during static initialization and the process is
aborted with an error message if we have compiled with a feature that is
not supported by the current CPU.

Such a check is already performed during static initialization of the
library `libtensorflow_framework.so.2`, but since this library performs
many actions during initialization, there is a high risk of triggering
an illegal instruction before we reach the check, depending on the
(undefined) static initialization order.

Because of this we compile a separate shared library that we load as a
self-check before loading `_pywrap_tensorflow_internal.so`.

We also change to use `<iostream>` for printing in case of a failed
check, as use of the logging framework might itself trigger an
illegal instruction (it did in my testing).

Tested by attempting to use a wheel built with AVX-512 on an AMD CPU:

Before:

```
$ python -c 'import tensorflow'
Illegal instruction
```

After:

```
$ python -c 'import tensorflow'
The TensorFlow library was compiled to use AVX512F instructions, but these aren't available on your machine.
Aborted
```