Retain device attribute if the op is placed on a TPU replicated core device
This device attribute is used during legalization to get the sharding of type MAXIMAL and the information is not redundant with the cluster attributes.

Note that the old bridge legalization handles sharding using this attribute but MLIR lowerings ignores this attribute. Added a TODO to use explicit sharding op from the front end.

PiperOrigin-RevId: 364911341
Change-Id: I4a7bf52feefcad856feb45af864931e03b74f87e