Revert logic to infer profiler_outdir from SummaryWriter
This reverts https://github.com/tensorflow/tensorflow/pull/35340 from Jan 2020. As pointed out in https://github.com/tensorflow/tensorboard/issues/4468, this change never worked (it has no tests). The bug is that that stashing the logdir on the SummaryWriter metadata results in an EagerTensor, when we need a Python str.

This is a revert rather than a fix forward because the current bug is symptomatic of larger issues with this approach. While in pure eager mode we can always evaluate the EagerTensor to get the string value, once tf.function is involved it's all more complicated. For example, if we were to allow building writers inside tf.functions, passing them out of the function would now require also passing the logdir together with the writer's resource tensor, which is a design constraint we'd prefer to avoid.

A better approach would be something avoids duplicating the logdir tracking in Python state alongside the actual underlying writer. This could be either A) more deeply integrating the profiler code with the summary system (so that it writes data via the writer resource itself), or alternatively B) adding a private API to retrieve the logdir of the current default writer (if it's logdir-based).

PiperOrigin-RevId: 365862669
Change-Id: Ida5d854cde4b0e4805aa77a8c992ac094f74cb60