Move `SummaryWriter.{set_,}as_default()` impls up to the base class
This anticipates some proliferation in the SummaryWriter class hierarchy by pulling the common logic around setting the writer as default up into the base class. This makes it slightly less of a pure interface, but the nature of these methods is such that they already are agnostic to the particular writer implementation (with the lone exception of `NoopSummaryWriter` which is a bit of a special case).

While I was at it, I've reworked `set_as_default()` to just call `as_default().__enter__()` to reduce duplication and the potential for divergence. This required defining the context manager as its own class versus just using `@contextmanager` for the reasons described in the code comment.

Lastly I fixed some docstrings to point to the base class instead of duplicating text.

PiperOrigin-RevId: 366139123
Change-Id: Ifa6d88ba2e87f1eaae96fa3ea2fe3f6c5b73234a