Merge the LowerCustomOps pass and the PrepareTf pass
This cl extracted the logic of the LowerCustomOps into an utility method so it
can be invoked the PrepareTf pass. Between the LowerCustomOps pass and the
PrepareTf pass, there is a shape inference pass and it will not work with the
tfl.TfCustomOp if it doesn't have a shape inference trait. This pass added a
simple shape inference for the wrapped tf.FakeQuant* ops.

PiperOrigin-RevId: 366820440
Change-Id: Ic72af7f62cc559d6e2cf2e79e2ed0a74f0d49379