Fix inspect_checkpoint tool crashing when printing variants.
This would happen when the `--all_tensors` flag was used.  We were throwing
an exception about being unable to convert the type to a numpy dtype and
exiting immediately.  This did not show a traceback, which made it unclear as
to whether the error was actually an error or was an intended part of
execution.  Any tensors that remained to be printed would not be, making it
seem like the tensors that had already been printed were the only ones
found in the checkpoint.

PiperOrigin-RevId: 367229703
Change-Id: I14e858fcc77e4219fd4baf366059549693440228