[TF:TRT] Avoid calling IExecutionContext::setDeviceMemory with nullptr when the device memory is released.
Previously, we call IExecutionContext::setDeviceMemory with nullptr when we
release the device memory for an execution context. The NVIDIA document says
that calling IExecutionContext::setDeviceMemory with nullptr is only acceptable
when the required memory for the context is zero. This change remove such a
call. We will follow up with NVIDIA for a way to represent that the device
memory for the context is not available.

PiperOrigin-RevId: 367455485
Change-Id: I9f8a739082bf6e8aec911234f67153dbe7740c1b